Rewrite main parser module as class and rename
